requirements.txt
# 光伏CAD智能设计系统依赖

# 核心机器学习
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.26.0
numpy>=1.21.0
scipy>=1.7.0

# CAD处理
ezdxf>=1.0.0
python-dotenv>=0.19.0

# 图像处理
opencv-python>=4.5.0
Pillow>=8.0.0
matplotlib>=3.5.0

# Web框架
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.0.0

# 数据处理
pandas>=1.3.0
jsonschema>=4.0.0

# 工具
tqdm>=4.60.0
requests>=2.25.0
pyyaml>=6.0.0